part-3-python-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "appreg"
version = "0.1.0"
description = "Application registry with environment-specific configuration overrides"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
appreg = "appreg.cli:cli"

[tool.setuptools.packages.find]
include = ["appreg*"]
